Respect days parameter when listing timeline events
Up until now, when the user requested hourly report (by selecting days
selector in the UI), the code always returned only the events from the
last day of selected interval.

This commit fixes this by unifying the time interval construction for
all types of reports, getting rid of the buggy date construction in
the process.