Stop N+1 with EmsDashboardService#recent_records
The addition of #recent_records in the following PR:

    https://github.com/ManageIQ/manageiq-ui-classic/pull/3427

Introduced the following N+1 in the code:

* * *

The code would create an initial query of the following:

    SELECT "vms".*
    FROM "vms"
    WHERE "vms"."type" IN ('MiqTemplate', ...)
      AND "vms"."template"= 't'
      AND (created_on > '2018-07-10 20:47:40.460291' and ems_id = 1)

This means it fetches all of the `vms` table, which is 88 columns, for a
given EMS and date range.

From this is triggered by the `records.sort_by(&:created_on)` in the
previous code, since `sort_by` is just inherited from `Enumerable` in
Ruby's stdlib.  It then calls a `.uniq` on the records, and then
iterates over each record (I am pretty sure this `uniq` does not do
anything, but can't be sure).

In the code, we `strftime` then applied on to the only column that was
actually needed for every row returned (besides the id), and then a
subsequent query is executed on each record to get every record that
matches that date (forgetting to do the original sort by `ems.id` as
well)

All to return a hash that looks like this:

    {
      "2018-01-01" => 123,
      "2018-01-02" => 234,
      "2018-01-03" => 345,
    }

* * *

The code change does all of this, but in SQL, only returning from the DB
an array of 2 element arrays of the same hash as above, and then
converts that array into a hash.

This is done in a single query then, and only instantiates enough data
to then generate the hash, but 0 `ActiveRecord` objects are necessary.