Use a more idiomatic way of building strings
String interpolation is the much preferred way to join strings in Ruby,
and also avoids issues were `:+` can sometimes not work as expected.

Worth noting:  I was getting errors on this line when running the code
locally when `@ems_id` was an integer.  It might be that in the
controller, it is passed in as a string, but this way, either works and
Ruby will convert it as needed.