Lazy load StorageManagers upon volume creation
When creating a new volume user has to pick what StorageManager is volume
supposed to be created with, hence we perform API request to fetch those.
Problem is we share angular controller with other volume operations (attach,
detach) that don't need this list of all managers and fetching those
anyway results in permission denied error due to fine-grained permission
setting.

For example if we give user only permission to attach/detach volume, but
not to view StorageManagers, then we must not perform the API call as it
results in 403 with red popup.

With this commit we refactor controller so that it now only performs the
API calls that are really needed.

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1571224

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>