Show correct VMs upon Service retirement
With this commit we make sure correct VMs are shown as part
of service retirement request. Talking about "Affected VMs"
section.

Prior to this commit there was always only a single VM displayed
in there, and it was not the right one. Happened to see AWS EC2 VM
on a vCloud vApp retirement request details!

The reason for all this was that

```
@miq_request.options[:src_ids]
```

were treated as it's a list of VM ids whereas it infact was a
list of Service ids. If we were lucky enough, some Service id
collided with random VM id and BUUUM, wrong VM was shown.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1632239

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>