rename 'container_metrics' tab to 'metrics'
'container_metrics' is actually a proper name for a tab to differentiate it from the RHV metrics tab.
So ideally  'container_metrics' tab name should be retained, however this would mean a whole bunch of other changes
from the backend side, since the backend expects cred_type to be 'metrics' (as opposed to 'container_metrics')

Therefore, unfortunately we have to rename the tab to 'metrics' . This is probably an easier change than changing the backend rails controller code.