Fixed feature id when checking assert_privileges.
Code was checking for feature id as ems_infra_tag when trying to tag selected storages thru Provider relationship screen. This was causing an issue when user did not have access to ems_infra_tag feature but had access to storage_tag feature. Fixed code to check if `@display` is set to use that when setting feature id to check for.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1610957