Allow a maximum of 500 characters in a service dialog text box in stead of 50
The service dialog text box is too short. This is bothering people for years now: https://github.com/ManageIQ/manageiq/issues/1226.
The best solution would be to add a 'maxlength' option to the dialog editor but the hereby proposed solution would make it workeable and would prevent people using a Text Area where they would prefer a text Box just because of the enforced maxlength of 50 characters for the text Box.