Use v_total_vms for HostDecorator
Instead of using `vms.size`, use `v_total_vms`, which will do the same
thing, unless the attribute exists on the model, in which it will avoid
making a call to the database.

This also means that the call for the size can be done in a single query
for an entire collection, instead of a N+1 manner.  This is already done
in views like `Host.yaml`, where `v_total_vms` is already in the `cols`.