set_ownership - use miq_template (not vm) for images
Right now, clicking the Set Ownership button on a nested list of images under a cloud provider redirects to `/vm/ownership`.

But image is not a vm, it's a template, so the url should be `/miq_template/ownership`.

The difference causes 2 bugs:

* no form buttons are shown
* affected items shows No records found

Fixing to use miq_template.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1650507