Fix missing paging in provisioning template selection
go to `/miq_request/pre_prov?template_klass=cloud`

there is a GTL with a list of templates,
but thanks to `@in_a_form = true` (and `center_div_no_listnav` checking both `layout_uses_paging? && !@in_a_form`),
that means no paging toolbar.

That was correct when this was not using the GTL component but no longer, removing.