AR::QueryMethods uniq/distinct takes no block
This never worked.  The argument passed to uniq/distinct is used to
either add or remove the 'distinct' constraint. Fortunately, the bug just
caused it to be inefficient.

Since AR::QueryMethods uniq is deprecated in 5.0 and removed in 5.1 and we
want the Array#uniq method, we need to convert the relation to an array.