pf-line-chart: fix errors when no data available
The pfLineChart component does deal with the possibility that `vm.chartData.dataAvailable` is false, and shows an empty chart instead.

However, before getting there, it fails, because there is no `vm.config` object when no data is available.

All of updateAll only deals with the situation where there is data available, so just bailing early when it isn't.