Set lazy loading in options instead of locals in TreeBuilders
The tree_init_options has a :lazy option that is used to decide
whether the node children should be included or not when building.
The set_locals_for_render contains an :autoload options that is
being used by the JS to decide to do autoloading or no. This
change is trying to unify the two behaviors into a single one.

I found some trees that had both settings but with different
boolean values, in those cases the :lazy option was taken as
reference because it has the higher precedence.

Also the false value is no longer necessary for the :lazy option
so I removed all those as well.