Container trends chart - wait for loadingDone before displaying pf-trends-chart
prevents

    TypeError: Cannot read property 'layout' of undefined
    at trendsChartController.e.updateAll (application-d966831b…581bed47155710.js:3)
    at trendsChartController.e.$onChanges (application-d966831b…581bed47155710.js:3)

when the JSON endpoints are taking longer to respond.

The controller already has a concept of loadingDone, used to display a spinner.
But it was never used to prevent rendering of the pf-trends-chart component, which fails when config is undefined.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1684226