miqSetToolbarCount - fix setting toolbar counts
miqSetButtons is not only called for toolbars, but also fake toolbars,
so, we can't always update the toolbar count when nonzero.

But for specific cases we can - so splitting the functionality into miqSetToolbarCount,
which will always set the count for toolbars.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1694712