Schedules spec - fix sporadic failure
Fixes:

    Failures:
     1) ReportController#schedule_edit reset rbac testing
        Failure/Error: expect(expected_id).to eq(schedule.id)
          expected: 66000000000038
               got: 66000000000039
          (compared using ==)
        # ./spec/controllers/report_controller/schedules_spec.rb:33:in `block (3 levels) in <top (required)>'

this is happening because resetting a form means calling `checked_or_params` to get the id of the schedule being editted, but the function was mocked to always give all the schedules.

So the sporadic failure was happening when the right schedule was second in the list, instead of first.