Use leaf subclasses with assets
Physical Infra Manager doesn't have assets as it's a pseudo abstract
class.

Fixes rails 5.1 deprecations such as:

DEPRECATION WARNING: The asset "svg/vendor-physical_infra_manager.svg" is not present in the asset pipeline.Falling back to an asset that may be in the public folder.
This behavior is deprecated and will be removed.
To bypass the asset pipeline and preserve this behavior,
use the `skip_pipeline: true` option.
 (called from quadrant_output at /Users/joerafaniello/.gem/ruby/2.4.6/bundler/gems/manageiq-decorators-fbb4695da1fb/app/helpers/quadicon_helper.rb:98)

DEPRECATION WARNING: The asset "svg/vendor-physical_infra_manager.svg" is not present in the asset pipeline.Falling back to an asset that may be in the public folder.
This behavior is deprecated and will be removed.
To bypass the asset pipeline and preserve this behavior,
use the `skip_pipeline: true` option.
 (called from post_process_textual_summary at /Users/joerafaniello/Code/manageiq-ui-classic/app/helpers/textual_summary_helper.rb:40)