Extend toolbar pluggability to all plugins.
Newly any plugin can add content to to toolbars. Not only providers.

Plugged-in toolbars should implement class method:

def self.record_valid?(record)
    ...
end

that decides if the toolbar is to be displayed for a particular record.

Example display only for RHEV VMs:

def self.record_valid?(record)
  record.kind_of?(ManageIQ::Providers::Vmware::InfraManager::Vm)
end

Please note that such level of control is only available for the
toolbars on the detail pages. Not for the toolbars on the listing pages.