Addressed PR feedback
When initially loading the form  wait for all the fields, and zones API request to finish, and then turn the spinner off. When action type is changed, it will do the 2 requests in parallel, and wait for requests to complete and then turn spinner off.
Fixed view to show Filters drop down for non-automate task schedule that broke with previous commit in this PR.
Fixed sort on Zone API call.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1415106