vm_common#reload spec - prevent sporadic failure when folder id matches vm id
the controller code is probably subtly wrong, but I'm not sure *how*.

If the id param should always win, maybe `params[:id].present? ? "v-#{params[:id]}" : x_node` would be the best solution in `reload`.
If `x_node` should win, it's not clear when.

But, when both the folder id and the vm id are the same, the code does not behave as expected.
Fixing in test for now.

(seed to reproduce sporadic failure: 47803)