count_only_or_object now works with ordered query
When the order is passed into `count_only_or_object` as the 3rd param,
the order is not passed to the count queries. So it works.

When the order is part of the query itself, it becomes part of the count
query, and this causes problems.

The fix is to remove the order from the count query. The result is the
same, but the query is faster and has fewer problems.

`responds_to?` is used to keep arrays working.

https://bugzilla.redhat.com/show_bug.cgi?id=1726790