Dialog copy - remove ids pointing to the original dialog
When copying a dialog, we load the original, allow the user to change it, and save that data as a new dialog.

But, the API just saves the fields, groups and tabs, then saves the dialog, and only after that updates those tabs to point to the new dialog. No transactions.
Which means that when dialog validation fails, all the copied tabs still point to the original dialog, leading to duplicate tabs appearing in the original.

This does not fix the API bug or attempt to wrap it in a transaction,
this just makes sure the UI won't send any obsolete ids when copying,
thus limiting the impact of the API bug to extra unused tabs remaining in the db.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1713100