Dialog editor - show validation errors
Right now, we never actually show dialog editor validation errors.
We're trying to show it on the submit button in ui-classic, but the button is disabled (and thus not rendering title) when validation fails.

So, moving to an explicit error div.

(related to https://bugzilla.redhat.com/show_bug.cgi?id=1729265)