Report data: add display of active filter. Add empty state.
The active filter display allows one click clearing of the filters as
requested by a BZ.

https://bugzilla.redhat.com/show_bug.cgi?id=1739264

The display of active filter is partly ready for the multi-criteria
filtering, once it's available in the API.

When there are no records, no grid and no paginator is displayed.
Instead an empty state component with a message is displayed.

https://bugzilla.redhat.com/show_bug.cgi?id=1740055