Optimization - spec for breadcrumbs when report.name and saved.name are identical
breadcrumbs deduplicate breadcrumbs with identical titles,
but for Optimization, the "saved reports under report" can have the same title as the "saved report detail" screen,
because the names of the saved reports are generated from the name of the report definition.

Adding a (failing) spec to make sure these don't get deduplicated in those conditions.

https://bugzilla.redhat.com/show_bug.cgi?id=1741203