Prevent login redirect with ?timeout during logout
When logging out via the Logout button,
when there are multiple HTTP requests happening,
the logout can cause API requests to fail with a 401 before the browser redirects to the login screen,
causing the response handler to call redirectLogin, to redirec to the login screen with a "session timed out" message.

We should not be claiming a timeout *during* logout, adding a check for that.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1741283