miqCheckForChanges - don't assume formReducer=={} means pristine=false
formReducer is always registered, so, we end up with formReducer == {} even on dashboard, where there are no forms.

Thus, miqCheckForChanges assumes this is a react form with .pristine == false (because undefined), and tries to show the "Abandon changes" popup.

Extending the reducer to also maintain a `in_a_form` boolean, and using that.