Toolbar: fix update of toolbar when starting from an empty one.
Previously when an empty toobar was rendered, listener for RX messages
was not installed (listenToRx). That resulted in a situation that a
toolbar update w/o a re-render of the toolbar <div> would not have any
affect.

This was visible in schedules under settings where toolbar would be
missing in the page.

Newly the listenToRx call is in the wrapping component (MiqToolbar) and
is installed no matter what toolbar content is passed in.