Fix SVG upload, render
In generic object, we allow upload of custom svg images, but the code only matches the old image/svg,
but can't handle image/svg+xml which is what chrome now uses.

Similarly, PicturesController sets the type to image/svg on download, which causes the browser to render a placeholder instead.
Fixing to use svg+xml for mime type, but keep svg as extension.