MiqFormRenderer: fix proptype error
```
Warning: Failed prop type: The prop `formFieldsMapper.function () { [native code] }` is marked as required in `MiqFormRenderer`, but its value is `undefined`.
    in MiqFormRenderer (created by ConnectFunction)
    in ConnectFunction (created by WorkersForm)
    in div (created by Grid)
    in Grid (created by WorkersForm)
    in WorkersForm
    in Provider
```

introduced in #6769.

No point in validating `formFieldsMapper` in `MiqFormRenderer`, we're just passing this unaltered to DDF FormRender, that's where that validation should happen.

The bug is that all Object fields are strings, so `{ [PropTypes.string]: 123 }` becomes `{ "function () { [native code] }": 123 }`,
that's not a valid proptype.