Webpack - remove settings we no longer need, fix production build
stats.errorDetails defaults to true,
output.pathinfo defaults to true,
optimization.minimize defaults to true in production mode,
CompressionPlugin algorithm defaults to gzip,
CompressionPlugin.asset was renamed to filename, but defaults to [path].gz[query] as of 2.0.

The last broke schema validation for the production settings, removing the asset param fixes it.