[Rails52] Use arel to fix "Dangerous query method"
There were multiple places (all in `.order` calls) that triggered the
following warning

  DEPRECATION WARNING: Dangerous query method (method whose arguments
  are used as raw SQL) called with non-attribute argument(s):
  "lower(name)".  Non-attribute arguments will be disallowed in Rails
  6.0. This method should not be called with user-provided values, such
  as request parameters or model attributes. Known-safe values can be
  passed by wrapping them in Arel.sql().

This fixes that by instead calling to `arel_table[].lower`, which does
the same thing, but uses the existing arel interface to do it, which
should be safer.  It should fail earlier for non-name columns, and
should be safer overall then using `Arel.sql()`.