[Gemfile] Remove "second" source
Since we inherit the `Gemfile` from `ManageIQ/manageiq` in this repo,
there is now a deprecation warning when running `bundler` v2.1.4:

    [DEPRECATED] Your Gemfile contains multiple primary sources. Using
    `source` more than once without a block is a security risk...

This is because there is a duplicate source when evaling the `Gemfile`
from `ManageIQ/manageiq`.  Since that one needs to stay, removing this
one allows `bundle update` and friends to continue to work without the
deprecation warning.