Display title on nested Policy Simulation screen.
Go to Cloud Provider summary screen. Click on Instances link, then from the list select one or more Instance(s), Press Policy Simulation button for one or more selected instance(s). On the Policy Simulation screen, Add a Policy Profile and then click on one of the Instances to see results of Policy simulation - Notice missing title on screen and an extra listnav on the left side. This fix removes the extra layout from the screen and adds the title on the Policy simulation result screen. Title was missing when going to the screen thru relationships(non-explorer style screen)