[cypress/commands.js] Add "wait" for menuItems
By using a `cy.get` with this command, it forces a `wait` for the DOM to
load with the menu.  Using `cy.window` doesn't (seem to) do that.

This probably could be smaller, but this is a workaround to fix the
currently breaking tests using this method (currently only
`menu.spec.js`).