[search_box.spec.js] Fix failing specs
For both specs, they suffered from the menu changes
ManageIQ/manageiq-ui-classic PR 6963, so those have been addressed to
use some form of an equivalent.

For the first spec ("is present"), it also had a random "div click" in
there that either doesn't make sense any more, is entirely unnecessary,
or there is a disconnect between the author of this commit (me) and the
original author of the spec.  Regardless, commenting out the line in
question allows the spec to pass.