[cypress/support/index.js] No screenshots on CI
Disables screenshots when running in travis.

cypress.env.json generation
---------------------------

Cypress handles environment variables itself, so `process.env` isn't
available.  As a result, "Option 2" is what is being done from here:

https://docs.cypress.io/guides/guides/environment-variables.html#Option-2-cypress-env-json

which makes sure a default config file exists as part of the `setup`
step for running the specs.  A `file` task is used here to only run it
if the file doesn't exist, but makes it a dependency for the `:setup`
task.

By default, screenshots are disabled in CI, but enabled in dev.

The cypress.env.json file is also ignored by git, so local edits can
happen and can be customized for each developer.  On CI, it will just be
copied in and the defaults from `.cypress.ci.env.json` will be what is
configured in that environment