Policy deletion - use ajaxReload instead of treeSelect to trigger reload with `@report_deleted=true`
`reload` calls `tree_select` with the just deleted node id,
which makes `valid_active_node` trigger a tree reload and switch to root

So using that path instead of directly tree_selecting root to trigger tree reload.

But it also causes an error flash because the record no longer exists,
which can be overriden by `@report_deleted`, so setting it in `reload` when `?deleted=true`.

And updated RemoveGenericItemModal to allow for ajaxReload=true in addition to treeSelect