Made changes to Edit event's action screen for a selected Policy
Made changes to Edit event's actions for a policy form, added a new drop down for user's to select an Event before they can change the actions for the Event. Previously this drop down was not needed because Events were listed as child nodes in the Policy tree on the left side.

https://github.com/ManageIQ/manageiq-ui-classic/issues/6996
https://github.com/ManageIQ/manageiq-ui-classic/issues/6819