[catalog_controller_spec.rb] Fix ems let... again
This is a follow up to 02e4b85 as the solution provided there, while
works for this use case, ended up causing problems in
`ManageIQ/manageiq-content`:

https://github.com/ManageIQ/manageiq/pull/20787#issuecomment-726357421

Instead, we use the :provider_embedded_ansible factory, and reference
the `.automation_manager` from that.  In addition, since a update to the
`:inventory_root_groups` was needed, we also modify the record in a
`before` statement as well.

This ends up not being as clean as the previous implementation, but
makes sure we are using the `EmbeddedAnsible` provider in the form that
it would be called when used in the application.