[spec] Fix create_user_with_group errors
When used in conjunction with the :user factory, if this is the first
spec to be run that creates a user, or the sequence for FactoryBot is
somehow reset (unlikely), this can cause a error with the `userid` field
since they are case sensitive.  Meaning the following:

- "User1"
- "user1"

Are considered the same id.  The format for `FactoryBot.create(:user)`
(defined in ManageIQ/manageiq) is `"user#{s}"`, where `s` is the current
sequence number, so this has a non-deterministic failure rate when using
a Random seed.

However, when checking out the commit prior to this one and running the
following:

  $ bundle exec rspec --seed 25104 spec/controllers/report_controller_spec.rb

The error is reproduced every single time.