[CI] Force bundler 2.2.9 (avoid 2.2.10)
For now, our bundler relies on dependent gems from rubygems.org for gems
that are in rubygems.manageiq.org.  This is because we use
rubygems.manageiq.org to host gems we have overwritten.

However, bundler 2.2.10 prevents this for security reasons, but doesn't
offer a way to handle our use case.  For now, sticking to bundler 2.2.9
should allow CI to continue to run.