[ops_controller/settings/zones.rb] Soft delete auth
Removes Auth from both the form, and the controller in the UI.  Still
allows them to be added via that `rails console`, but removes the
feature from the UI, with the intent to possibly remove from the product
all together in a future release.

Note:  Documentation should probably also be updated to remove
mentioning this.