Handle queue workers inheriting values from worker_base.
The existing code was checking only the worker's value and falling back to
queue_worker_base but not falling back to the worker_base if the other two
locations had no value.

This causes the select's value to be the first item in the select, which is 200 MB.
If you as a user try to change some other unrelated value and then save, all
of these "incorrect" 200 MB values get saved instead of the correct value
inherited from worker_base.

For example, smart_proxy_worker was correct because that worker class specified a
memory_threshold.  Reporting, C & U processor and collector, Generic, and
Priority workers were all showing 200 MB in the select because that's the first
entry in the select and because the default of 1.gigabyte was coming from
worker_base and no value was coming from queue_worker_base or the worker class
itself.

For more information, review https://github.com/ManageIQ/manageiq/blob/19b57d3c88679b12718e301f4a19d4b1fee11705/config/settings.yml#L1063-L1113 and see that the reporting worker specifies no memory_threshold and the queue_worker_base also specifies no memory_threshold but should inherit the worker_base value of 1.gigabytes.  This is the same problem for the other workers mentioned above.