Fix not yet renamed chargeback_rates features
Previously, you'll get a permission error when trying to "save" or "reset" when
adding or editing a chargeback rate.

chargeback_rates_edit is not yet chargeback_rate_edit
and chargeback_rates_new is not yet chargeback_rate_new.

It looks like this was broken in 968c4657af3b414e262ad5499dbdaf345445ed5d
where the "new" renamed features were used in a few places but the product
feature has not yet been renamed.

Changed the tests to use these features instead of blindly getting all features
in order to recreate the problem.