[ApplicationController::Tags] Fix bulk_reassignment
Fixes editing of tags, where instead of passing in the proper adds and
deletes, all tags were passed in as deletes and adds, and only because
adding comes after deleting did this work previously, but at the cost of
performance.

By ensuring we are doing array subtraction with the same types, it fixes
the `:add_ids` and `:delete_ids` from being incorrect and causing a
surge of ActiveRecord queries when it isn't necessary.