Add app/views/dark.scss
Adds a `dark.scss` theme to the application stylesheets to allow for a
future dark theme (and others, if desired) to be added to MIQ.

Uses the $carbon--theme--g90 for the theme and uses the "overwrite"
approach for handling the styles.

A future commit could switch to `css` variables to allow for different
ways of switching the theme, but for now this is a decent start, and a
future commit can hook it up to be enabled via user setting.