[WIP] Add user setting for ui theme
Technically this setting key was already used:

  app/controllers/configuration_controller.rb:425

but this enables the setting to be used to set the class to enable the
dark styles based on user preference.

The form just uses the existing User settings, and adds a `theme` key to
the DDF form to allow this to be configured when rendering the page.
The key is then used in a helper to set a CSS class in the layout,
enabling the theme.