update dashboard layout after removing a widget
Before:

The widget was getting removed from the WidgetSet (aka dashboard)
But the blob which contained further information was not updated.
So the widget was still getting displayed

After:

Updating the user dashboard so it no longer contained the widget
Pass in the widget so it did not need to be looked up a second time.