Prevent drag drop widgets with dashboard_view role
Users where able to drag and drop widgets in the dashboard regardless if they had the dashboard view or add role.
When users with only view permissions tried to do this, an error would appear on screen which told them they didn't have
the correct permissions. With this change users with only the view rights will not be able to move widgets.