Fix strings not in message catalogs due to a bug in parsing []
Previously, this string[1]:
was not detected as a gettext string to be translated.  This is because the
gettext_i18n_rails_js parser regexp [2] does not handle some javascript syntax
that could be used within the __(), n_(), etc. method calls.  Specifically,
[]+-*/ used outside of ``, "", or '' but within the n_() could fail the regexp
and not be detected as a string to be translated.  It's possible this could also
happen in N_(), s_(), p_(), variants, but it's unclear if this is used and may
be unlikely to occur.

Clearly, this is a bug in the parser and we should try to upstream any fix but
since it's not really using a javascript parser, it's probably best to avoid doing
anything but very basic javascript in arguments to these methods.

Here is an example from irb:

```
irb(main):007:0> require 'gettext_i18n_rails_js'
irb(main):008:0> `cat test.js`
=> "n__('vm', 'vms', provider[0].vms.size)\n"
irb(main):009:0> puts GettextI18nRailsJs::Parser::Javascript.parse('./test.js').inspect
[]
=> nil
irb(main):010:0> `cat test2.js`
=> "n__('vm', 'vms', vms.size)\n"
irb(main):011:0> puts GettextI18nRailsJs::Parser::Javascript.parse('./test2.js').inspect
[["vm\u0000vms", "./test2.js:1"]]
```

Note, how the second example doesn't use a [] in the value used to determine the
plural form and the parser detects the strings vm and vms and the file location.
The first example has [] there and the parser fails to detect any strings.

After making this change, and running bundle exec rake locale:update_all, we now
have this added to the english po and pot:

```
@ locale/en/manageiq.po:31731 @ msgstr ""
 msgid "File is empty"
 msgstr ""

+#: ../app/javascript/forms/mappers/validatorMapper.jsx:21
+msgid "File is too large, maximum allowed size is %s bytes. Current file has %s byte"
+msgid_plural "File is too large, maximum allowed size is %s bytes. Current file has %s bytes"
+msgstr[0] ""
+msgstr[1] ""
```

[1] https://github.com/ManageIQ/manageiq-ui-classic/blob/39e7bdb81bf35dffb0201603bab2508b9a66c6fe/app/javascript/forms/mappers/validatorMapper.jsx#L20-L24
[2] https://github.com/webhippie/gettext_i18n_rails_js/blob/v1.3.1/lib/gettext_i18n_rails_js/parser/javascript.rb#L90-L129