Use lazy loaded default uglifier
Part of #8300

Previously, we eager loaded uglifier to provide a custom configuration.

This was added in https://github.com/ManageIQ/manageiq/pull/9878

In #8300, we verified we no longer need this custom configuration.
Additionally, if we use symbolic :uglifier, Uglifier won't be eager loaded,
execjs won't be loaded from uglifier, and therefore node wouldn't be needed.

This should allow us to only need node when we compile assets at build time.